Switch to devmode to solve a few issues
This commit takes the controversial step of using "devmode"
confinement instead of "strict" confinement.

While there is no doubt that "strict" confinement is theoretically
preferable, "devmode" confinement gives us the same structure
($PATH points to binaries within the snap) but with access to the wider
system should we need it.

(Read about the confinement options here:
https://snapcraft.io/docs/snap-confinement)

"strict" confinement offers far greater security, but we ran into a
couple of insurmountable issues with strict confinement:

- https://forum.snapcraft.io/t/access-to-users-for-dotrun-snap/15459
- https://forum.snapcraft.io/t/python-multiprocessing-permission-denied-in-strictly-confined-snap/15518

The first issue we can work around by mounting the `/User` directory
for mac users into somewhere inside `/home/ubuntu` in the multipass VM.
This isn't ideal, as any paths that are output in logs will be confusing
for the mac users, but it will work.

The second issue, though, is a real blocker. We need to be able to run
multiprocessing successfully. Particularly, we need to be able to run
black successfully, but access to multiprocessing in general is pretty
important.

Given that the audience for this tool is developers in our team, I'm
not sure it matters to much to ask them to run `dotrun` in `devmode` -
hopefully we can ask them to trust this snap and our projects -
and so for now it seems to me to make the most sense to simply ask
everyone to install dotrun with `--beta --devmode` flags.

We also don't need any plugs with devmode as all access is available
anyway.

I've also found a way to make the .venv python virtualenv persist
successfully through upgrades of the dotrun snap,
so it no longer tracks the snap revision in `.dotrun.json`, or
regenerates the python env if it encounters a new snap version.