Update the mechanism for getting the IP address
This updates the recommendation in the README for how to
tell the local development server which hostname to listen at.

The important detail is that if dotrun is run within multipass
then the server will be run using the IP address of that multipass VM,
so that in the server's logs you'll see e.g.
"Running on http://10.3.4.5" and be able to click on or copy that
address directly to access the server.

This should also work if dotrun is run on the host machine, where it
will probably show something like `http://192.168.1.12` or again
`http://10.77.44.22`.