Remove advice to run server with public IP
Stop recommending that we use `hostname -I | awk '{print $1;}'` to get
the public IP for the server to listen at. This came out of this PR:
https://github.com/canonical-web-and-design/juju.is/pull/34#discussion_r392619167

The reason we were suggesting this was so that the log would output
a server address that would work even if the server is running inside
multipass. Without this, the user has to know how to find the
IP address of their multipass instance and manually type it into
the address bar to see the running server, rather than clicking on or
copying the server address that is output in the logs.

However, the new recommendations for running dotrun inside multipass
now require the user to have some understanding of how multipass works,
so also understanding how to find the point to the IP address seems a
less cumbersome requirement.

There were two downsides of the `hostname` approach:

- That command won't work in macOS, so anyone trying to run our
  `package.json` script directly on a macOS system without `dotrun`
  would fail, which isn't very nice
- It can lead to the server being run only in a public IP, and so not
  being available on localhost. E.g. if we run
  `gunicorn --bind 192.168.1.2:8111`, then the server will only be
  available at http://192.168.1.2:8111, and not http://localhost:8111,
  http://127.0.0.1:8111 or http://0.0.0.0:8111 as the developer might
  reasonable expect. This could be pretty frustrating